chore(functions): use proper Java test naming
